修复了判断字符串函数因为使用了无符号类型导致的bug，优化了代码。

+++ vtos/src/base/os_string.c
@@ -150,7 +150,7 @@ os_size_t os_str_find(const char *str, const char *pattern)
 	return ret;
 }
 
-int32 os_str_cmp(const char *str1, const char *str2)
+int8 os_str_cmp(const char *str1, const char *str2)
 {
 	for (; *str1 != '\0' && *str2 != '\0'; str1++, str2++)
 	{
